Add date_trunc to scalar functions to support date_trunc during ingestion (#6538)
* Support date_trunc during ingestion

* Fix test and make case insensitive
