+++ pinot-common/src/main/resources/zone-index.properties
