+++ pinot-core/src/main/java/org/apache/pinot/core/operator/transform/function/DateTruncTransformFunction.java
@@ -22,23 +22,13 @@ import com.google.common.base.Preconditions;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
+import org.apache.pinot.common.function.DateTimeUtils;
+import org.apache.pinot.common.function.TimeZoneKey;
 import org.apache.pinot.core.common.DataSource;
 import org.apache.pinot.core.operator.blocks.ProjectionBlock;
 import org.apache.pinot.core.operator.transform.TransformResultMetadata;
 import org.apache.pinot.core.plan.DocIdSetPlanNode;
-import org.joda.time.Chronology;
 import org.joda.time.DateTimeField;
-import org.joda.time.DateTimeFieldType;
-import org.joda.time.DateTimeZone;
-import org.joda.time.DurationField;
-import org.joda.time.DurationFieldType;
-import org.joda.time.chrono.ISOChronology;
-import org.joda.time.field.DividedDateTimeField;
-import org.joda.time.field.OffsetDateTimeField;
-import org.joda.time.field.ScaledDurationField;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 
 /**
  * The <code>DateTruncTransformationFunction</code> class implements the sql compatible date_trunc function for TIMESTAMP type.
@@ -90,38 +80,12 @@ public class DateTruncTransformFunction extends BaseTransformFunction {
   public static final String EXAMPLE_INVOCATION =
       String.format("%s('week', time_expression, 'seconds', <TZ>, <Output-Granularity>)", FUNCTION_NAME);
   private static final String UTC_TZ = TimeZoneKey.UTC_KEY.getId();
-  private static final DateTimeFieldType QUARTER_OF_YEAR = new QuarterOfYearDateTimeField();
-  private static final Logger LOGGER = LoggerFactory.getLogger(DateTruncTransformFunction.class);
   private TransformFunction _mainTransformFunction;
   private TransformResultMetadata _resultMetadata;
   private long[] _longOutputTimes;
   private DateTimeField _field;
   private TimeUnit _inputTimeUnit;
   private TimeUnit _outputTimeUnit;
-
-  private static DateTimeField getTimestampField(ISOChronology chronology, String unitString) {
-    switch (unitString) {
-      case "millisecond":
-        return chronology.millisOfSecond();
-      case "second":
-        return chronology.secondOfMinute();
-      case "minute":
-        return chronology.minuteOfHour();
-      case "hour":
-        return chronology.hourOfDay();
-      case "day":
-        return chronology.dayOfMonth();
-      case "week":
-        return chronology.weekOfWeekyear();
-      case "month":
-        return chronology.monthOfYear();
-      case "quarter":
-        return QUARTER_OF_YEAR.getField(chronology);
-      case "year":
-        return chronology.year();
-    }
-    throw new IllegalArgumentException("'" + unitString + "' is not a valid Timestamp field");
-  }
 
   @Override
   public String getName() {
@@ -147,7 +111,7 @@ public class DateTruncTransformFunction extends BaseTransformFunction {
             : inputTimeUnitStr;
     TimeZoneKey timeZoneKey = TimeZoneKey.getTimeZoneKey(timeZone);
 
-    _field = getTimestampField(DateTimeZoneIndex.getChronology(timeZoneKey), unit);
+    _field = DateTimeUtils.getTimestampField(DateTimeUtils.getChronology(timeZoneKey), unit);
     _resultMetadata = LONG_SV_NO_DICTIONARY_METADATA;
     _outputTimeUnit = TimeUnit.valueOf(outputTimeUnitStr);
   }
@@ -170,87 +134,5 @@ public class DateTruncTransformFunction extends BaseTransformFunction {
           .convert(_field.roundFloor(TimeUnit.MILLISECONDS.convert(input[i], _inputTimeUnit)), TimeUnit.MILLISECONDS);
     }
     return _longOutputTimes;
-  }
-
-  public static final class DateTimeZoneIndex {
-    private static final DateTimeZone[] DATE_TIME_ZONES;
-    private static final ISOChronology[] CHRONOLOGIES;
-    private static final int[] FIXED_ZONE_OFFSET;
-    private static final int VARIABLE_ZONE = Integer.MAX_VALUE;
-
-    private DateTimeZoneIndex() {
-    }
-
-    public static ISOChronology getChronology(TimeZoneKey zoneKey) {
-      return CHRONOLOGIES[zoneKey.getKey()];
-    }
-
-    public static DateTimeZone getDateTimeZone(TimeZoneKey zoneKey) {
-      return DATE_TIME_ZONES[zoneKey.getKey()];
-    }
-
-    static {
-      DATE_TIME_ZONES = new DateTimeZone[TimeZoneKey.MAX_TIME_ZONE_KEY + 1];
-      CHRONOLOGIES = new ISOChronology[TimeZoneKey.MAX_TIME_ZONE_KEY + 1];
-      FIXED_ZONE_OFFSET = new int[TimeZoneKey.MAX_TIME_ZONE_KEY + 1];
-      for (TimeZoneKey timeZoneKey : TimeZoneKey.getTimeZoneKeys()) {
-        short zoneKey = timeZoneKey.getKey();
-        DateTimeZone dateTimeZone;
-        try {
-          dateTimeZone = DateTimeZone.forID(timeZoneKey.getId());
-        } catch (IllegalArgumentException e) {
-          LOGGER.error("Exception while extracting time zone field", e);
-          continue;
-        }
-        DATE_TIME_ZONES[zoneKey] = dateTimeZone;
-        CHRONOLOGIES[zoneKey] = ISOChronology.getInstance(dateTimeZone);
-        if (dateTimeZone.isFixed() && dateTimeZone.getOffset(0) % 60_000 == 0) {
-          FIXED_ZONE_OFFSET[zoneKey] = dateTimeZone.getOffset(0) / 60_000;
-        } else {
-          FIXED_ZONE_OFFSET[zoneKey] = VARIABLE_ZONE;
-        }
-      }
-    }
-  }
-
-  // Original comment from presto code
-  // ```Forked from org.elasticsearch.common.joda.Joda```
-  public static final class QuarterOfYearDateTimeField extends DateTimeFieldType {
-    private static final DateTimeFieldType QUARTER_OF_YEAR = new QuarterOfYearDateTimeField();
-    private static final long serialVersionUID = -5677872459807379123L;
-    private static final DurationFieldType QUARTER_OF_YEAR_DURATION_FIELD_TYPE = new QuarterOfYearDurationFieldType();
-
-    private QuarterOfYearDateTimeField() {
-      super("quarterOfYear");
-    }
-
-    @Override
-    public DurationFieldType getDurationType() {
-      return QUARTER_OF_YEAR_DURATION_FIELD_TYPE;
-    }
-
-    @Override
-    public DurationFieldType getRangeDurationType() {
-      return DurationFieldType.years();
-    }
-
-    @Override
-    public DateTimeField getField(Chronology chronology) {
-      return new OffsetDateTimeField(
-          new DividedDateTimeField(new OffsetDateTimeField(chronology.monthOfYear(), -1), QUARTER_OF_YEAR, 3), 1);
-    }
-
-    private static class QuarterOfYearDurationFieldType extends DurationFieldType {
-      private static final long serialVersionUID = -8167713675442491871L;
-
-      public QuarterOfYearDurationFieldType() {
-        super("quarters");
-      }
-
-      @Override
-      public DurationField getField(Chronology chronology) {
-        return new ScaledDurationField(chronology.months(), QUARTER_OF_YEAR_DURATION_FIELD_TYPE, 3);
-      }
-    }
   }
 }
